[chore/TK-141]: SharedPaymentCategoryClassificationService 분리
