+++ src/main/java/withbeetravel/service/payment/SharedPaymentCategoryClassificationService.java
@@ -1,0 +1,8 @@
+package withbeetravel.service.payment;
+
+import withbeetravel.domain.Category;
+
+public interface SharedPaymentCategoryClassificationService {
+
+    public Category getCategory(String storeName);
+}
